requirements.txt
fastapi[standard]==0.115.5
uvicorn==0.32.0
python-dotenv==0.19.0
pydantic==2.9.2
cobo-waas2==1.6.0
web3==7.4.0
PyJWT==2.8.0
cryptography==43.0.3
